requirements.txt
requests
beautifulsoup4
google-generativeai
tqdm 
python-dotenv 